refactor: 현재 url과 비교 감지하는 로직 커스텀 훅으로 분리
